Result template changes
Updating the filenames of the output files to match what they are currently called, just to be able to get further in the processing. The longer-term solution is obviously to decide on certain names so that we always know what to expect. At this time, only the first pixel is processed.